pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ffaa"
version = "0.1.0"
description = "Multimodal Large Language Model based Explainable Open-World Face Forgery Analysis Assistant."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "torch==2.1.2", "torchvision==0.16.2",
    "transformers==4.37.2", "tokenizers==0.15.1", "sentencepiece==0.1.99", "shortuuid",
    "accelerate==0.21.0", "peft", "bitsandbytes",
    "pydantic", "markdown2[all]", "numpy", "scikit-learn==1.2.2",
    "gradio==4.16.0", "gradio_client==0.8.1",
    "requests", "httpx==0.24.0", "uvicorn", "fastapi",
    "einops==0.6.1", "einops-exts==0.0.4", "timm==0.6.13",
    "albumentations==1.4.11", "dlib==19.24.6", "grad_cam==1.5.2",
    "huggingface_hub", "opencv_python_headless==4.10.0.84", 
    "pillow==10.4.0",
]

[project.optional-dependencies]
train = ["deepspeed==0.12.6", "flash_attn==2.5.8", "ninja", "wandb"]
build = ["build", "twine"]

[project.urls]
"Homepage" = "https://ffaa-vl.github.io/"
"Bug Tracker" = "https://github.com/thu-huangzc/FFAA/issues"

[tool.setuptools.packages.find]
exclude = ["benchmark*", "checkpoints*", "docs", "experiments", "heatmaps*", "playground*", "scripts*"]

[tool.wheel]
exclude = ["benchmark*", "checkpoints*", "docs", "experiments", "heatmaps*", "playground*", "scripts*"]